requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
oauthlib==3.0.1
Pillow==6.2.1
PySocks==1.6.8
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
tweepy==3.7.0
urllib3==1.24.3
pymongo==3.9.0
